[spark] override equals as a preparative step for performance optimization
